more commands lots of changes process or rewriting neural network implementation restructured and cleaned logging framework

+++ src/main/java/gk646/jnet/neuralnetwork/builder/ActivationFunction.java
@@ -4,17 +4,17 @@ package gk646.jnet.neuralnetwork.builder;
 public enum ActivationFunction {
     SIGMOID {
         @Override
-        public float apply(float x) {
-            return (float) (1 / (1 + Math.exp(-x)));
+        public double apply(double x) {
+            return  (1 / (1 + Math.exp(-x)));
         }
     },
     RELU {
         @Override
-        public float apply(float x) {
+        public double apply(double x) {
             return Math.max(0, x);
         }
     };
 
 
-    public abstract float apply(float x);
+    public abstract double apply(double x);
 }
